[level 1] Title: 콜라츠 추측, Time: 0.05 ms, Memory: 73 MB -BaekjoonHub

+++ 프로그래머스/1/12943. 콜라츠 추측/콜라츠 추측.java
@@ -1,5 +1,5 @@
 class Solution {
-    public int solution(double num) {
+    public int solution(long num) {
         int cnt = 0;
         
         while (num != 1) {
